Combine AccountAccess structs

+++ crates/cheatcodes/spec/src/vm.rs
@@ -204,7 +204,7 @@ interface Vm {
         /// An ordered list of storage accesses made during an account access operation.
         StorageAccess[] storageAccesses;
         /// Call depth traversed during the recording of state differences
-        uint256 depth;
+        uint64 depth;
     }
 
     /// The storage accessed during an `AccountAccess`.
